refactor: RacingGame 모델 분리
